move definition provider under document provider
