+++ src/cloudWatchLogs/activation.ts
@@ -18,23 +18,20 @@ import { LogStreamRegistry } from './registry/logStreamRegistry'
 import { Commands } from '../shared/vscode/commands2'
 import { searchLogGroup } from './commands/searchLogGroup'
 import { changeLogSearchParams } from './changeLogSearch'
-import { JumpToStream } from './commands/searchLogGroup'
 
 export async function activate(context: vscode.ExtensionContext, configuration: Settings): Promise<void> {
     const settings = new CloudWatchLogsSettings(configuration)
     const registry = new LogStreamRegistry(settings)
 
-    const definitionProvider = new JumpToStream(registry)
+    const documentProvider = new LogStreamDocumentProvider(registry)
+
     vscode.languages.registerDefinitionProvider(
-        { language: 'log', scheme: CLOUDWATCH_LOGS_SCHEME, pattern: '.*:.+' },
-        definitionProvider
+        { language: 'log', scheme: CLOUDWATCH_LOGS_SCHEME, pattern: '**' },
+        documentProvider
     )
 
     context.subscriptions.push(
-        vscode.workspace.registerTextDocumentContentProvider(
-            CLOUDWATCH_LOGS_SCHEME,
-            new LogStreamDocumentProvider(registry)
-        )
+        vscode.workspace.registerTextDocumentContentProvider(CLOUDWATCH_LOGS_SCHEME, documentProvider)
     )
 
     context.subscriptions.push(
