+++ src/cloudWatchLogs/commands/searchLogGroup.ts
@@ -186,48 +186,3 @@ export class SearchLogGroupWizard extends Wizard<SearchLogGroupWizardResponse> {
         this.form.timeRange.bindPrompter(() => new TimeFilterSubmenu())
     }
 }
-
-export class JumpToStream implements vscode.DefinitionProvider {
-    private registry: LogStreamRegistry
-
-    public constructor(registry: LogStreamRegistry) {
-        this.registry = registry
-    }
-    async provideDefinition(
-        document: vscode.TextDocument,
-        position: vscode.Position,
-        token: vscode.CancellationToken
-    ): Promise<vscode.Definition | vscode.LocationLink[] | undefined> {
-        const activeUri = document.uri
-        const logGroupInfo = parseCloudWatchLogsUri(activeUri).logGroupInfo
-        const curLine = document.lineAt(position.line)
-        let streamUri: vscode.Uri
-        console.log('here')
-        try {
-            const streamIDMap = this.registry.getStreamIdMap(activeUri)
-            const streamID = streamIDMap?.get(curLine.lineNumber)
-            if (!streamID) {
-                throw new Error(`cwl: No streamID found for stream with uri ${activeUri}`)
-            }
-            const parameters: CloudWatchLogsParameters = {
-                limit: this.registry.configuration.get('limit', 10000),
-            }
-            logGroupInfo.streamName = streamID
-            streamUri = createURIFromArgs(logGroupInfo, parameters)
-            const initialStreamData: CloudWatchLogsData = {
-                data: [],
-                parameters: parameters,
-                logGroupInfo: logGroupInfo,
-                retrieveLogsFunction: getLogEventsFromUriComponents,
-                busy: false,
-            }
-            await this.registry.registerLog(streamUri, initialStreamData)
-            // Set the document language
-            const doc = await vscode.workspace.openTextDocument(streamUri)
-            vscode.languages.setTextDocumentLanguage(doc, 'log')
-            return new vscode.Location(streamUri, new vscode.Position(0, 0))
-        } catch (err) {
-            throw new Error(`cwl: Error determining definition for content in ${document.fileName}`)
-        }
-    }
-}
